stunprobe: compare the two address lengths as one type
socklen_t is int on mingw-w64 while getaddrinfo's ai_addrlen is size_t
there, so the de-duplication that skips a repeated record compared a
signed with an unsigned type and the Windows build stopped on
-Werror=sign-compare. glibc has both unsigned, which is why the native
build and both VM builds were clean.

Compare them as size_t. The mingw-w64 cross compiler on this host
reproduces the failure and accepts the fix.

Fixes: 5131cef90bb9 ("stunprobe, peering: one probe path for both families")

+++ src/stunprobe.c
@@ -348,7 +348,7 @@ static int resolve_name(const char *server, int family,
 		    (size_t)ai->ai_addrlen > sizeof(out[0]))
 			continue;
 		for (i = 0; i < n; i++)	/* getaddrinfo may repeat one */
-			if (outlen[i] == ai->ai_addrlen &&
+			if ((size_t)outlen[i] == (size_t)ai->ai_addrlen &&
 			    !memcmp(&out[i], ai->ai_addr, (size_t)outlen[i]))
 				break;
 		if (i < n)
